Unify exec_utils.py and utils.py
