+++ libsubmit/providers/cobalt/cobalt.py
@@ -4,7 +4,7 @@ import time
 from string import Template
 
 import libsubmit.error as ep_error
-from libsubmit.exec_utils import wtime_to_minutes
+from libsubmit.utils import wtime_to_minutes
 from libsubmit.launchers import launchers
 from libsubmit.providers.cobalt.template import template_string
 from libsubmit.providers.provider_base import ExecutionProvider
